Move consumer into a Python package
This adds a pyproject.toml to define a new Python package and fixes the
code and tests for the adjusted imports. We want everything to be in a
Python package so that it's easy to use the fedora-messaging
command-line interface to manage the consumer and its configuration.
