+++ pyproject.toml
@@ -1,0 +1,41 @@
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[project]
+name = "fedora-cloud-tests"
+description = "An AMQP message consumer that automates the validation of Fedora nightly cloud images."
+readme = "README.md"
+license = "MIT"
+license-files = { paths = ["LICENSE"] }
+dynamic = ["version"]
+requires-python = ">=3.10"
+
+dependencies = [
+    "fedora-messaging",
+    "fedora-image-uploader-messages",
+    "lisa[azure] @ git+https://github.com/microsoft/lisa.git@20250819.1",
+]
+
+test = [
+    "pytest",
+    "pytest-asyncio"
+]
+
+[tool.hatch.version]
+path = "fedora_cloud_tests/__init__.py"
+
+[tool.hatch.metadata]
+# Required for LISA since they don't publish releases to PyPI
+allow-direct-references = true
+
+[tool.black]
+line-length = 100
+
+[tool.isort]
+profile = "black"
+
+[tool.coverage.run]
+source = [
+    "fedora_cloud_tests/",
+]
